# apps/api/spec/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "acting-api",
    "description" : "연기 분석 플랫폼 API v2. Bearer access token을 사용합니다.",
    "version" : "0.1.0"
  },
  "paths" : {
...
    "/v2/push-tokens" : {
      "post" : {
        "tags" : [ "v2-push" ],
        "summary" : "Register Push Token",
        "description" : "이 단말의 Expo push token 을 현재 회원 것으로 등록한다. 멱등. 보호 기능이라 동의와 프로필이\n끝난 회원만 부를 수 있다 — 동의 전에는 기기 정보를 받지 않는다. 게스트는 403 member_only.",
        "operationId" : "register_push_token_v2_push_tokens_post",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterPushTokenRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Successful Response"
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "v2-push" ],
        "summary" : "Unregister Push Token",
        "description" : "이 단말의 토큰을 지운다. 로그아웃의 첫 단계와, 로그아웃 때 실패한 삭제의 다음 실행 재시도에서\n부른다. 멱등. <b>로그인 없이 받는다</b> — 푸시 토큰을 갖고 있다는 것이 본인 확인이고, 주인을\n따지지 않고 그 토큰 행을 지운다. 한 IP 에서 분당 60회까지다.",
        "operationId" : "unregister_push_token_v2_push_tokens_delete",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UnregisterPushTokenRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Successful Response"
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/guest/transfer-code" : {
      "post" : {
        "tags" : [ "v2-guest" ],
        "summary" : "Issue Transfer Code",
        "description" : "웹이 보여 줄 여섯 자리 숫자를 만든다. 10분 유효, 한 번 쓰면 끝, 새로 받으면 이전 코드는\n무효다. 게스트만 부를 수 있고 필요한 동의 문서는 없다.",
        "operationId" : "issue_transfer_code_v2_guest_transfer_code_post",
        "responses" : {
          "201" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransferCodeResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
    "/v2/guest-transfers" : {
      "post" : {
        "tags" : [ "v2-guest" ],
        "summary" : "Transfer Guest Data",
        "description" : "앱에서 코드를 넣으면 그 게스트의 자료 행의 주인을 회원으로 바꾼다. 한 트랜잭션이고, 끝나면\n게스트 계정을 닫는다. 회원과 게스트 둘 다 배우 기억이 있는데 memory_choice 가 없으면 아무것도\n옮기지 않고 409 다 — 코드는 살아 있다.",
        "operationId" : "transfer_guest_v2_guest_transfers_post",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GuestTransferRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful Response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuestTransferResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Validation Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security" : [ {
          "HTTPBearer" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RegisterPushTokenRequest" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "token" : {
            "type" : "string",
            "maxLength" : 512,
            "minLength" : 1,
            "title" : "Token"
          },
          "platform" : {
            "type" : "string",
            "enum" : [ "ios", "android" ],
            "title" : "Platform"
          }
        },
        "required" : [ "platform", "token" ],
        "title" : "RegisterPushTokenRequest"
      },
...
      "TransferCodeResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "code" : {
            "type" : "string",
            "title" : "Code"
          },
          "expires_in" : {
            "type" : "integer",
            "title" : "Expires In"
          },
          "expires_at" : {
            "type" : "string",
            "format" : "date-time",
            "title" : "Expires At"
          }
        },
        "required" : [ "code", "expires_at", "expires_in" ],
        "title" : "TransferCodeResponse"
      },
      "GuestTransferRequest" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "code" : {
            "type" : "string",
            "title" : "Code"
          },
          "memory_choice" : {
            "$ref" : "#/components/schemas/MemoryChoice"
          }
        },
        "required" : [ "code" ],
        "title" : "GuestTransferRequest"
      },
      "MemoryChoice" : {
        "type" : "string",
        "enum" : [ "member", "guest" ],
        "title" : "MemoryChoice"
      },
      "GuestTransferResponse" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "transferred" : {
            "type" : "boolean",
            "title" : "Transferred"
          }
        },
        "required" : [ "transferred" ],
        "title" : "GuestTransferResponse"
      },
...
      "HTTPValidationError" : {
        "type" : "object",
        "properties" : {
          "detail" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ValidationError"
            },
            "title" : "Detail"
          }
        },
        "title" : "HTTPValidationError"
      },
      "ValidationError" : {
        "type" : "object",
        "properties" : {
          "loc" : {
            "type" : "array",
            "items" : {
              "anyOf" : [ {
                "type" : "string"
              }, {
                "type" : "integer"
              } ]
            },
            "title" : "Location"
          },
          "msg" : {
            "type" : "string",
            "title" : "Message"
          },
          "type" : {
            "type" : "string",
            "title" : "Error Type"
          },
          "input" : {
            "title" : "Input"
          },
          "ctx" : {
            "type" : "object",
            "title" : "Context"
          }
        },
        "required" : [ "loc", "msg", "type" ],
        "title" : "ValidationError"
      },
...
      "UnregisterPushTokenRequest" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "token" : {
            "type" : "string",
            "maxLength" : 512,
            "minLength" : 1,
            "title" : "Token"
          }
        },
        "required" : [ "token" ],
        "title" : "UnregisterPushTokenRequest"
      },
...
    },
    "securitySchemes" : {
      "HTTPBearer" : {
        "type" : "http",
        "scheme" : "bearer"
      }
    }
  }
}